Split run tests
